requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
Werkzeug==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
confluent-kafka==2.3.0